Slide Tracking Backend
