+++ Database/MySQL/BlockCountReports/qryBlockCountByTimes.sql
@@ -42,7 +42,7 @@ SELECT subTblSlides.BlockID AS subBlockID, subTblSlideDistribution.SlideDistribu
         GROUP BY subTblSlides.BlockID, SlideDistributionLocation, subTblSlideDistribution.DTReadyForCourier
 
 /*qryFirstRunBlockCount*/
-SELECT Count(qrySubBlockCountWLocation.subBlockID) AS BlockCount, SlideDistributionLocation
+SELECT Count(qrySubBlockCountWLocation.subBlockID) AS FirstRunBlockCount, SlideDistributionLocation
 FROM (SELECT subTblSlides.BlockID AS subBlockID, subTblSlideDistribution.SlideDistributionLocation
         FROM tblSlides as subTblSlides
         INNER JOIN   tblSlideDistribution as subTblSlideDistribution on subTblSlides.SlideDistributionID = subTblSlideDistribution.SlideDistributionID
@@ -52,7 +52,7 @@ FROM (SELECT subTblSlides.BlockID AS subBlockID, subTblSlideDistribution.SlideDi
 GROUP BY SlideDistributionLocation;
 
 /*qrySecondRunBlockCount*/
-SELECT Count(qrySubBlockCountWLocation.subBlockID) AS BlockCount, SlideDistributionLocation
+SELECT Count(qrySubBlockCountWLocation.subBlockID) AS SecondRunBlockCount, SlideDistributionLocation
 FROM (SELECT subTblSlides.BlockID AS subBlockID, subTblSlideDistribution.SlideDistributionLocation
         FROM tblSlides as subTblSlides
         INNER JOIN   tblSlideDistribution as subTblSlideDistribution on subTblSlides.SlideDistributionID = subTblSlideDistribution.SlideDistributionID
@@ -60,3 +60,40 @@ FROM (SELECT subTblSlides.BlockID AS subBlockID, subTblSlideDistribution.SlideDi
          AND  subTblSlideDistribution.DTReadyForCourier < funCurrentDaySecondRunCutoff()
         GROUP BY subTblSlides.BlockID, SlideDistributionLocation) as qrySubBlockCountWLocation
 GROUP BY SlideDistributionLocation;
+
+/*qryThirdRunBlockCount*/
+SELECT Count(qrySubBlockCountWLocation.subBlockID) AS ThirdRunBlockCount, SlideDistributionLocation
+FROM (SELECT subTblSlides.BlockID AS subBlockID, subTblSlideDistribution.SlideDistributionLocation
+        FROM tblSlides as subTblSlides
+        INNER JOIN   tblSlideDistribution as subTblSlideDistribution on subTblSlides.SlideDistributionID = subTblSlideDistribution.SlideDistributionID
+        WHERE subTblSlideDistribution.DTReadyForCourier >= funCurrentDaySecondRunCutoff()
+         AND  subTblSlideDistribution.DTReadyForCourier < funCurrentDayThirdRunCutoff()
+        GROUP BY subTblSlides.BlockID, SlideDistributionLocation) as qrySubBlockCountWLocation
+GROUP BY SlideDistributionLocation;
+
+/*qryFourthRunBlockCount*/
+SELECT Count(qrySubBlockCountWLocation.subBlockID) AS FourthRunBlockCount, SlideDistributionLocation
+FROM (SELECT subTblSlides.BlockID AS subBlockID, subTblSlideDistribution.SlideDistributionLocation
+        FROM tblSlides as subTblSlides
+        INNER JOIN   tblSlideDistribution as subTblSlideDistribution on subTblSlides.SlideDistributionID = subTblSlideDistribution.SlideDistributionID
+        WHERE subTblSlideDistribution.DTReadyForCourier >= funCurrentDayThirdRunCutoff()
+         AND  subTblSlideDistribution.DTReadyForCourier < funCurrentDayFourthRunCutoff()
+        GROUP BY subTblSlides.BlockID, SlideDistributionLocation) as qrySubBlockCountWLocation
+GROUP BY SlideDistributionLocation;
+
+/*qryBlockCountAllRunTimesBySortVal*/
+SELECT 
+    `OPENLIS`.`tblSlideDistributionLocations`.`LocAbbr` AS `LocAbbr`,
+    `qryFirstRunBlockCount`.`FirstRunBlockCount` AS `FirstRunBlockCount`,
+    `qrySecondRunBlockCount`.`SecondRunBlockCount` AS `SecondRunBlockCount`,
+    `qryThirdRunBlockCount`.`ThirdRunBlockCount` AS `ThirdRunBlockCount`,
+    `qryFourthRunBlockCount`.`FourthRunBlockCount` AS `FourthRunBlockCount`,
+    `qryAllRunsBlockCount`.`TotalBlockCount` AS `TotalBlockCount`
+FROM
+    (((((`OPENLIS`.`tblSlideDistributionLocations`
+    LEFT JOIN `OPENLIS`.`qryFirstRunBlockCount` ON ((`qryFirstRunBlockCount`.`SlideDistributionLocation` = `OPENLIS`.`tblSlideDistributionLocations`.`LocationID`)))
+    LEFT JOIN `OPENLIS`.`qrySecondRunBlockCount` ON ((`OPENLIS`.`tblSlideDistributionLocations`.`LocationID` = `qrySecondRunBlockCount`.`SlideDistributionLocation`)))
+    LEFT JOIN `OPENLIS`.`qryThirdRunBlockCount` ON ((`OPENLIS`.`tblSlideDistributionLocations`.`LocationID` = `qryThirdRunBlockCount`.`SlideDistributionLocation`)))
+    LEFT JOIN `OPENLIS`.`qryFourthRunBlockCount` ON ((`OPENLIS`.`tblSlideDistributionLocations`.`LocationID` = `qryFourthRunBlockCount`.`SlideDistributionLocation`)))
+    LEFT JOIN `OPENLIS`.`qryAllRunsBlockCount` ON ((`OPENLIS`.`tblSlideDistributionLocations`.`LocationID` = `qryAllRunsBlockCount`.`SlideDistributionLocation`)))
+ORDER BY `OPENLIS`.`tblSlideDistributionLocations`.`SortValue`
